remove legacy functions
and move to mc_expect_grad the only function that was still used.
